GPT-SoVITS-beta/requirements.txt
numpy==1.23.4
scipy
tensorboard
librosa==0.9.2
numba==0.56.4
pytorch-lightning
gradio>=4.0,<=4.24.0
ffmpeg-python
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
tqdm==4.66.6
funasr==1.0.27
cn2an
pypinyin
pyopenjtalk>=0.3.4
g2p_en
torchaudio
modelscope==1.10.0
sentencepiece
transformers==4.46.1
chardet
PyYAML
psutil
jieba_fast
jieba
LangSegment>=0.2.0
Faster_Whisper
wordsegment
rotary_embedding_torch
pyjyutping
g2pk2
ko_pron
opencc; sys_platform != 'linux'
opencc==1.1.6; sys_platform == 'linux'
python_mecab_ko; sys_platform != 'win32'
fastapi<0.112.2
pyjyutping


pymysql==1.1.1
DBUtils==3.1.0
soundfile==0.12.1
apscheduler==3.10.4

torch==2.0.0  --index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.1  --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.0  --index-url https://download.pytorch.org/whl/cu118